add Remote Attestation to Scenario #3
by setting up remote attestation for Spark's driver and executor pods,
we ensure that only the authorized code is getting any secrets or
sensitive configuration, even in an untrusted host, such as a virtual
machine in the cloud.

we can additionally check the platform's TCB or require an external
service, such as Microsoft Azure Attestation, to also verify the
applications.